修复了新浪微博登录的问题

+++ website/cas/src/main/java/com/uhuila/website/cas/oauth/sinaweibo/SinaWeiboProvider.java
@@ -65,7 +65,7 @@ public class SinaWeiboProvider extends AbstractOAuth20Provider {
         JsonNode subJson = userJson.get("id");
         if (subJson != null) {
             OAuthUserProfileHelper
-                    .addAttribute(userProfile, "uid", subJson.getIntValue());
+                    .addAttribute(userProfile, "uid", subJson.getLongValue());
         }
         subJson = userJson.get("domain");
         if (subJson != null) {
